# backend/opd/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chettinad Care v2 \u2014 Connected OPD API",
    "version": "2.0.0",
    "description": "Implemented Connected OPD REST contract. API URL retains /api/v1 for compatibility; v2 workflows live under /opd. All /opd routes require a role-bound bearer token and active account. Patient access is limited to self, doctors to linked care, nurses to their department. Mutable resources use __v; conflicts return HTTP 409 and require refetch. Timestamps are UTC ISO 8601; schedule dates and times use Asia/Kolkata. See docs/V2_ARCHITECTURE.md for adapter and deployment requirements."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Same-origin deployment or Vite development proxy"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/opd/session": {
      "get": {
        "summary": "Read current role and profile link",
        "operationId": "get_opd_session",
        "tags": [
          "Authentication"
        ],
        "description": "Read current role and profile link",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "default": {
            "description": "Structured validation, authorization, conflict or internal error. Common statuses: 401, 403, 404, 409, 422, 429, 503.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-roles": [
          "PATIENT",
          "ADMIN",
          "NURSE",
          "DOCTOR"
        ]
      }
    },
    "/opd/directory": {
      "get": {
        "summary": "List departments, doctors and schedules",
        "operationId": "get_opd_directory",
        "tags": [
          "Scheduling"
        ],
        "description": "List departments, doctors and schedules",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Directory"
                }
              }
            }
          },
          "default": {
            "description": "Structured validation, authorization, conflict or internal error. Common statuses: 401, 403, 404, 409, 422, 429, 503.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-roles": [
          "PATIENT",
          "ADMIN",
          "NURSE",
          "DOCTOR"
        ]
      }
    },
    "/opd/slots": {
      "get": {
        "summary": "List available appointment slots",
        "operationId": "get_opd_slots",
        "tags": [
          "Scheduling"
        ],
        "description": "Asia/Kolkata date, up to 181 days ahead. Excludes past slots, breaks, unavailable doctors and overlapping bookings.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Slot"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Structured validation, authorization, conflict or internal error. Common statuses: 401, 403, 404, 409, 422, 429, 503.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "doctor_id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "x-roles": [
          "PATIENT",
          "ADMIN",
          "NURSE",
          "DOCTOR"
        ]
      }
    },
    "/opd/schedules": {
      "post": {
        "summary": "Create or update a doctor weekday schedule",
        "operationId": "post_opd_schedules",
        "tags": [
          "Scheduling"
        ],
        "description": "Create or update a doctor weekday schedule",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "default": {
            "description": "Structured validation, authorization, conflict or internal error. Common statuses: 401, 403, 404, 409, 422, 429, 503.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-roles": [
          "ADMIN"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleInput"
              }
            }
          }
        }
      }
    },
    "/opd/unavailability": {
      "get": {
        "summary": "List upcoming doctor leave",
        "operationId": "get_opd_unavailability",
        "tags": [
          "Scheduling"
        ],
        "description": "List upcoming doctor leave",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Unavailability"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Structured validation, authorization, conflict or internal error. Common statuses: 401, 403, 404, 409, 422, 429, 503.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-roles": [
          "ADMIN"
        ]
      },
      "post": {
        "summary": "Block a doctor date",
        "operationId": "post_opd_unavailability",
        "tags": [
          "Scheduling"
        ],
        "description": "Existing confirmed/checked-in appointments on the date must be rescheduled first.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "default": {
            "description": "Structured validation, authorization, conflict or internal error. Common statuses: 401, 403, 404, 409, 422, 429, 503.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-roles": [
          "ADMIN"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnavailabilityInput"
              }
            }
          }
        }
      }
    },
    "/opd/unavailability/{id}": {
      "delete": {
        "summary": "Restore doctor availability",
        "operationId": "delete_opd_unavailability_id",
        "tags": [
          "Scheduling"
        ],
        "description": "Restore doctor availability",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "default": {
            "description": "Structured validation, authorization, conflict or internal error. Common statuses: 401, 403, 404, 409, 422, 429, 503.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "x-roles": [
          "ADMIN"
        ]
      }
    },
    "/opd/patients": {
      "get": {
        "summary": "Search patient demographics",
        "operationId": "get_opd_patients",
        "tags": [
          "Patients"
        ],
        "description": "Matches name, mobile or MRN, maximum 100 results. Reception cannot access the clinical record endpoint.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Patient"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Structured validation, authorization, conflict or internal error. Common statuses: 401, 403, 404, 409, 422, 429, 503.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          }
        ],
        "x-roles": [
          "ADMIN"
        ]
      },
      "post": {
        "summary": "Register a demographic record",
        "operationId": "post_opd_patients",
        "tags": [
          "Patients"
        ],
        "description": "Issues a unique MRN and prevents duplicate phone registration. Does not create an encounter.",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "default": {
            "description": "Structured validation, authorization, conflict or internal error. Common statuses: 401, 403, 404, 409, 422, 429, 503.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-roles": [
          "ADMIN"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientProfile"
              }
            }
          }
        }
      }
    },
    "/opd/profile": {
      "get": {
        "summary": "Read own patient profile",
        "operationId": "get_opd_profile",
        "tags": [
          "Patients"
        ],
        "description": "Read own patient profile",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "default": {
            "description": "Structured validation, authorization, conflict or internal error. Common statuses: 401, 403, 404, 409, 422, 429, 503.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-roles": [
          "PATIENT"
        ]
      }
    },
    "/opd/patients/{id}": {
      "patch": {
        "summary": "Update authorized patient demographics",
        "operationId": "patch_opd_patients_id",
        "tags": [
          "Patients"
        ],
        "description": "Requires current patient __v. Mobile cannot change through this endpoint; it is the verified login identity.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "default": {
            "description": "Structured validation, authorization, conflict or internal error. Common statuses: 401, 403, 404, 409, 422, 429, 503.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "x-roles": [
          "ADMIN",
          "PATIENT"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientUpdate"
              }
            }
          }
        }
      }
    },
    "/opd/appointments": {
      "get": {
        "summary": "List role-scoped appointments",
        "operationId": "get_opd_appointments",
        "tags": [
          "Scheduling"
        ],
        "description": "Optional Asia/Kolkata date filter. Patients see own appointments; doctors their bookings; nurses their department. Maximum 500.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Appointment"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Structured validation, authorization, conflict or internal error. Common statuses: 401, 403, 404, 409, 422, 429, 503.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "x-roles": [
          "PATIENT",
          "ADMIN",
          "NURSE",
          "DOCTOR"
        ]
      },
      "post": {
        "summary": "Book an offered slot",
        "operationId": "post_opd_appointments",
        "tags": [
          "Scheduling"
        ],
        "description": "Transactions and doctor/patient locking prevent overlapping bookings. 409 indicates a slot conflict; refetch slots.",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          },
          "default": {
            "description": "Structured validation, authorization, conflict or internal error. Common statuses: 401, 403, 404, 409, 422, 429, 503.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-roles": [
          "ADMIN",
          "PATIENT"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Booking"
              }
            }
          }
        }
      }
    },
    "/opd/appointments/{id}/check-in": {
      "post": {
        "summary": "Verify identity and issue a queue token",
        "operationId": "post_opd_appointments_id_check_in",
        "tags": [
          "Queue"
        ],
        "description": "Only on appointment date in Asia/Kolkata. Creates one active encounter and department/day token. Repeating check-in returns the existing queue entry.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueEntry"
                }
              }
            }
          },
          "default": {
            "description": "Structured validation, authorization, conflict or internal error. Common statuses: 401, 403, 404, 409, 422, 429, 503.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "x-roles": [
          "ADMIN"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "identity_verified": {
                    "type": "boolean",
                    "const": true
                  }
                },
                "required": [
                  "identity_verified"
                ]
              }
            }
          }
        }
      }
    },
    "/opd/appointments/{id}/{action}": {
      "post": {
        "summary": "Reschedule, cancel or record no-show",
        "operationId": "post_opd_appointments_id_action",
        "tags": [
          "Scheduling"
        ],
        "description": "Reschedule, cancel or record no-show",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "default": {
            "description": "Structured validation, authorization, conflict or internal error. Common statuses: 401, 403, 404, 409, 422, 429, 503.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "path",
            "name": "action",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "reschedule",
                "cancel",
                "no-show"
              ]
            }
          }
        ],
        "x-roles": [
          "ADMIN",
          "PATIENT"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentAction"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          }
        },
        "required": [
          "success"
        ]
      },
...
      "Session": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "PATIENT",
              "ADMIN",
              "NURSE",
              "DOCTOR"
            ]
          },
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "patient_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ]
          },
          "must_change_password": {
            "oneOf": [
              {
                "type": "boolean"
              },
              {
                "type": "integer",
                "enum": [
                  0,
                  1
                ]
              }
            ]
          }
        }
      },
...
    }
  }
}